requirements.txt
python-whois==0.7.3
dnspython==2.4.0